+++ example/aws_cloudwatch_example.dart
@@ -9,8 +9,8 @@ const String ServiceInstance = 'Logs_XXXXXXXX';
 // Logging Variables
 const String LogGroup = 'LogGroupExample';
 String logStreamName = 'LogStreamExample';
-CloudWatch cloudWatch = new CloudWatch(
-    AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, Region, ServiceInstance);
+CloudWatch cloudWatch =
+    new CloudWatch(AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, Region);
 
 void log(String logString) {
   cloudWatch.log(logString);